+++ src/manifest.js
@@ -34,9 +34,8 @@ const contentScriptsDev = {
   matches: ["*://www.npmjs.com/package/*"],
   all_frames: true,
   js: [
-    "./scripts/lib/analytics.js",
     "./scripts/services/npmFaves-storage-service.js",
-    "./scripts/services/npmFaves-tracking-service.js",
+    "./scripts/services/npmFaves-analytics-service.js",
     "./scripts/utils/npmFaves-ui-notification.js",
     "./scripts/utils/npmFaves-helpers.js",
     "./scripts/content-script.js",
@@ -49,11 +48,7 @@ const nameProd = "npm faves";
 const contentScriptsProd = {
   matches: ["*://www.npmjs.com/package/*"],
   all_frames: true,
-  js: [
-    "./scripts/analytics.min.js",
-    "./scripts/npm-faves.core.min.js",
-    "./scripts/content-script.js",
-  ],
+  js: ["./scripts/npm-faves.core.min.js", "./scripts/content-script.js"],
   css: ["./styles/npm-faves.ui.min.css"],
 };